Update for issue #8
Added : "_ We also recommend to take backups of the database once a day or week, depending on the amount of data being consumed. In our case, it could've saved us the lost data. Since our project has stopped, we've not implemented an automatic backup script to avoid this in the future._"